Cleanup (IK doesn't compile now)
