+++ polkadot-parachains/common/src/lib.rs
@@ -25,6 +25,7 @@ pub mod fee;
 
 pub mod porteer;
 pub mod xcm_config;
+pub mod xcm_helpers;
 
 pub use constants::*;
 pub use types::*;
