+++ polkadot-parachains/common/src/porteer.rs
@@ -1,5 +1,23 @@
+// Copyright 2021 Integritee AG and Supercomputing Systems AG
+// This file is part of the "Integritee parachain" and is
+// based on Cumulus from Parity Technologies (UK) Ltd.
+
+// Integritee parachain is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Cumulus is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Integritee parachain.  If not, see <http://www.gnu.org/licenses/>.
+
+//! Common code that the Porteer requires to send code back and forth between IK<>IP.
+
 use crate::{AccountId, Balance};
-use frame_support::pallet_prelude::Weight;
 use pallet_porteer::XcmFeeParams;
 use parity_scale_codec::Encode;
 use sp_std::vec;
@@ -7,16 +25,14 @@ use xcm::{
 	latest::{
 		Asset, AssetFilter,
 		AssetTransferFilter::{ReserveDeposit},
-		Location, OriginKind, Parent, WeightLimit, WildAsset, Xcm,
+		Location, OriginKind, Parent, WildAsset, Xcm,
 	},
 	prelude::{
-		AllCounted, BurnAsset, BuyExecution, ClearOrigin, DepositAsset, ExecuteXcm, Fungible,
-		GlobalConsensus, Here, InitiateTransfer, Kusama, Parachain, PayFees, Polkadot,
-		ReceiveTeleportedAsset, RefundSurplus, SendXcm, SetAppendix, Transact, Wild, WithdrawAsset,
-		XcmError,
+		DepositAsset, Fungible,
+		GlobalConsensus, InitiateTransfer, Kusama, Parachain, PayFees, Polkadot,
+		ReceiveTeleportedAsset, RefundSurplus, SetAppendix, Transact, WithdrawAsset,
 	},
 };
-use xcm_executor::XcmExecutor;
 
 pub const IK_FEE: u128 = 1000000000000;
 pub const AHK_FEE: u128 = 33849094374679;
@@ -66,129 +82,15 @@ pub fn ip_cousin_v5() -> Location {
 	Location::new(2, [GlobalConsensus(Polkadot), Parachain(INTEGRITEE_POLKADOT_PARA_ID)])
 }
 
-pub fn asset_hub_kusama_location() -> Location {
+pub fn ahk_cousin_location() -> Location {
 	Location::new(2, [GlobalConsensus(Kusama), Parachain(ASSET_HUB_KUSAMA_PARA_ID)])
 }
 
-pub fn asset_hub_polkadot_location() -> Location {
+pub fn ahp_cousin_location() -> Location {
 	Location::new(2, [GlobalConsensus(Polkadot), Parachain(ASSET_HUB_POLKADOT_PARA_ID)])
 }
 
-/// Returns an Xcm that is meant to be executed locally to burn the native asset.
-pub fn burn_native_xcm<Call>(who: Location, amount: Balance, local_fees: Balance) -> Xcm<Call> {
-	let asset: Asset = (Here, Fungible(amount)).into();
-	burn_asset_xcm(who, asset, local_fees)
-}
-
-/// Returns an Xcm that is meant to be executed locally to burn the `Asset`.
-pub fn burn_asset_xcm<Call>(who: Location, asset: Asset, local_fees: Balance) -> Xcm<Call> {
-	Xcm(vec![
-		WithdrawAsset(vec![asset.clone(), (Here, Fungible(local_fees)).into()].into()),
-		PayFees { asset: (Here, Fungible(local_fees)).into() },
-		SetAppendix(Xcm(vec![
-			RefundSurplus,
-			DepositAsset { assets: AssetFilter::Wild(WildAsset::All), beneficiary: who },
-		])),
-		BurnAsset(asset.into()),
-	])
-}
-
-pub fn receive_teleported_asset<Call>(asset: Asset, beneficiary: Location) -> Xcm<Call> {
-	Xcm(vec![
-		ReceiveTeleportedAsset(asset.clone().into()),
-		ClearOrigin,
-		BuyExecution { fees: asset, weight_limit: WeightLimit::Unlimited },
-		DepositAsset { assets: Wild(AllCounted(1)), beneficiary },
-	])
-}
-
-pub fn teleport_asset<XcmConfig: xcm_executor::Config>(
-	who: Location,
-	beneficiary: Location,
-	asset: Asset,
-	local_fee: Balance,
-	destination: Location,
-) -> Result<(), XcmError> {
-	let xcm = burn_asset_xcm(who.clone(), asset.clone(), local_fee);
-	let remote_xcm = receive_teleported_asset(asset.into(), beneficiary);
-
-	execute_local_and_remote_xcm::<XcmConfig, <XcmConfig as xcm_executor::Config>::RuntimeCall>(
-		who,
-		xcm,
-		destination,
-		remote_xcm,
-	)?;
-	Ok(())
-}
-
-pub fn execute_local_and_remote_xcm<XcmConfig: xcm_executor::Config<RuntimeCall = Call>, Call>(
-	who: Location,
-	local_xcm: Xcm<Call>,
-	destination: Location,
-	remote_xcm: Xcm<()>,
-) -> Result<(), XcmError> {
-	let mut hash = local_xcm.using_encoded(sp_io::hashing::blake2_256);
-	let outcome = XcmExecutor::<XcmConfig>::prepare_and_execute(
-		who,
-		local_xcm,
-		&mut hash,
-		Weight::MAX,
-		Weight::zero(),
-	);
-
-	outcome.ensure_complete().map_err(|error| {
-		log::error!("Local execution is incomplete: {:?}", error);
-		error
-	})?;
-
-	let (ticket, _delivery_fees) = <XcmConfig as xcm_executor::Config>::XcmSender::validate(
-		&mut Some(destination),
-		&mut Some(remote_xcm),
-	)?;
-
-	<XcmConfig as xcm_executor::Config>::XcmSender::deliver(ticket)?;
-	Ok(())
-}
-
-/// XCM as it is being sent from the local Integritee chain to its cousin.
-pub fn local_integritee_xcm<Call, IntegriteePolkadotCall: Encode>(
-	call: IntegriteePolkadotCall,
-	local_fee: Balance,
-	local_as_sibling: Location,
-	local_as_cousin: Location,
-	ah_sibling: (Location, Balance),
-	ah_cousin: (Location, Balance),
-	integritee_cousin_as_sibling: (Location, Balance),
-) -> Xcm<Call> {
-	Xcm(vec![
-		// Assume that we always pay in native for now
-		WithdrawAsset((Here, Fungible(local_fee + ah_sibling.1)).into()),
-		PayFees { asset: (Here, Fungible(local_fee)).into() },
-		SetAppendix(Xcm(vec![
-			RefundSurplus,
-			DepositAsset { assets: AssetFilter::Wild(WildAsset::All), beneficiary: Here.into() },
-		])),
-		BurnAsset((Here, Fungible(ah_sibling.1)).into()),
-		// InitiateTransfer {
-		// 	destination: ah_sibling.0,
-		// 	remote_fees: Some(Teleport(AssetFilter::Definite(
-		// 		Asset { id: Here.into(), fun: Fungible(ah_sibling.1) }.into(),
-		// 	))),
-		// 	preserve_origin: true,
-		// 	assets: Default::default(),
-		// 	remote_xcm: ah_sibling_xcm(
-		// 		call,
-		// 		local_as_sibling,
-		// 		local_as_cousin,
-		// 		ah_cousin,
-		// 		integritee_cousin_as_sibling,
-		// 	),
-		// },
-	])
-}
-
-/// Nested XCM to be executed as `remote_xcm` from within `local_integritee_xcm` on the
-/// first hop, namely the Asset Hub sibling.
+/// XCM to be executed on the first hop, namely the Asset Hub sibling.
 pub fn ah_sibling_xcm<Call, IntegriteePolkadotCall: Encode>(
 	call: IntegriteePolkadotCall,
 	teleported_asset: Asset,
